+++ src/app/components/conversation-info/info-group/info-group.component.ts
@@ -70,18 +70,18 @@ export class InfoGroupComponent implements OnInit, AfterViewInit, OnChanges {
       .pipe(filter(event => event instanceof NavigationEnd))
 
       .subscribe((event: NavigationEnd) => {
-        console.log('InfoGroupComponent -  router.events prev url :', event.url);
+        // console.log('InfoGroupComponent -  router.events prev url :', event.url);
         this.previousUrl = event.url;
 
 
         this.unsubscribe$.next();
         this.unsubscribe$.complete();
         this.member_array = []
-        console.log('InfoGroupComponent -  router.events member_array :', this.member_array);
+        // console.log('InfoGroupComponent -  router.events member_array :', this.member_array);
         // this.pleaseDeleteMeEvent.emit();
 
-        if (this.groupDetail.hasOwnProperty("member_array")) {
-          console.log('InfoGroupComponent - router.events has Property member_array  :', this.groupDetail.hasOwnProperty("member_array"));
+        if (this.groupDetail && this.groupDetail.hasOwnProperty("member_array")) {
+          // console.log('InfoGroupComponent - router.events has Property member_array  :', this.groupDetail.hasOwnProperty("member_array"));
           delete this.groupDetail['member_array'];
         }
       });