+++ src/chat21-core/providers/firebase/firebase-conversations-handler.ts
@@ -282,8 +282,7 @@ export class FirebaseConversationsHandler extends ConversationsHandlerService {
                 this.logger.log('[FIREBASEConversationsHandlerSERVICE] conversationDetail childSnapshot.val() *****', childSnapshot.val());
                 this.logger.log('[FIREBASEConversationsHandlerSERVICE] conversationDetail childSnapshot *****', childSnapshot)
                 // && childData.uid
-                if (childSnapshot && childSnapshot.key) {
-                // if (childData) {
+                if (childSnapshot && childSnapshot.key && childData) {
                     childData.uid = childSnapshot.key;
                     const conversation = this.completeConversation(childData);
                     if (conversation) {
@@ -291,6 +290,7 @@ export class FirebaseConversationsHandler extends ConversationsHandlerService {
                     } else {
                         callback(null)
                     }
+
                 }
                 // this.BSConversationDetail.next(conversation);
             });
@@ -400,7 +400,7 @@ export class FirebaseConversationsHandler extends ConversationsHandlerService {
                 this.conversationAdded.next(conversationAdded);
             }
         } else {
-            this.logger.error('[FIREBASEConversationsHandlerSERVICE]ADDED::conversations with conversationId: ', childSnapshot.key, 'is not valid')
+            this.logger.log('[FIREBASEConversationsHandlerSERVICE]ADDED::conversations with conversationId: ', childSnapshot.key, 'is not valid')
         }
     }